+++ source/core/calibration/Parameter.cpp
@@ -1,4 +1,4 @@
-#include <core/calibration/parameter.h>
+#include <core/calibration/Parameter.h>
 #include <fmt/format.h>
 #include <core/util/compare.h>
